Fixed format specifier problems in v2 test suite -Also changed to use pmix_strncpy() in several spots
(cherry picked from commit c28f67e6d91fe6778272edbc7d8cd6637e0a7fd0)

+++ test/test_v2/test_common.h
@@ -75,7 +75,7 @@ extern FILE *pmixt_outfile;
     {                                                                                        \
         struct timeval tv;                                                                   \
         gettimeofday(&tv, NULL);                                                             \
-        double ts = tv.tv_sec + 1E-6 * tv.tv_usec;                                           \
+        double ts = (double)tv.tv_sec + 1E-6 * (double)tv.tv_usec;                           \
         fprintf(pmixt_outfile, "==%d== [%lf] %s:%s: %s\n", getpid(), ts, STRIPPED_FILE_NAME, \
                 __func__, pmix_test_output_prepare x);                                       \
         fflush(pmixt_outfile);                                                               \
@@ -94,7 +94,7 @@ extern FILE *pmixt_outfile;
     {                                                                                            \
         struct timeval tv;                                                                       \
         gettimeofday(&tv, NULL);                                                                 \
-        double ts = tv.tv_sec + 1E-6 * tv.tv_usec;                                               \
+        double ts = (double)tv.tv_sec + 1E-6 * (double)tv.tv_usec;                               \
         fprintf(stderr, "==%d== [%lf] ERROR [%s:%d:%s]: %s\n", getpid(), ts, STRIPPED_FILE_NAME, \
                 __LINE__, __func__, pmix_test_output_prepare x);                                 \
         fflush(stderr);                                                                          \
